refactor: move SenderSpecializedProcessorPort out of core (P5)
Replace the shared core port with adapter-internal interfaces:
BinanceSenderProcessor (adapter-binance) and CoinbaseSenderProcessor
(adapter-coinbase), both package-private. Internal routing mechanics
of an adapter are not a core concern.
