+++ adapter-binance/src/main/java/com/marmitt/binance/processor/send/BinanceSenderMessageProcessor.java
@@ -5,14 +5,14 @@ import com.marmitt.binance.BinanceUrlBuilder;
 import com.marmitt.binance.auth.BinanceRequestSigner;
 import com.marmitt.core.dto.websocket.request.MessageRequest;
 import com.marmitt.core.ports.outbound.exchange.adapter.SenderMessageProcessorPort;
-import com.marmitt.core.ports.outbound.exchange.adapter.SenderSpecializedProcessorPort;
+
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class BinanceSenderMessageProcessor implements SenderMessageProcessorPort {
 
-    private final List<SenderSpecializedProcessorPort> specializedProcessors;
+    private final List<BinanceSenderProcessor> specializedProcessors;
 
     public BinanceSenderMessageProcessor(ObjectMapper objectMapper, BinanceRequestSigner signer, BinanceUrlBuilder urlBuilder) {
         this.specializedProcessors = new ArrayList<>();
