+++ adapter-binance/src/main/java/com/marmitt/binance/processor/send/BinanceSenderProcessor.java
@@ -1,9 +1,9 @@
-package com.marmitt.core.ports.outbound.exchange.adapter;
+package com.marmitt.binance.processor.send;
 
 import com.marmitt.core.dto.websocket.request.MessageRequest;
 import com.marmitt.core.enums.MessageType;
 
-public interface SenderSpecializedProcessorPort {
+interface BinanceSenderProcessor {
 
     String execute(MessageRequest request);
