+++ adapter-coinbase/src/main/java/com/marmitt/coinbase/processor/send/CancelOrderProcessor.java
@@ -3,9 +3,9 @@ package com.marmitt.coinbase.processor.send;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.marmitt.core.dto.websocket.request.MessageRequest;
 import com.marmitt.core.enums.MessageType;
-import com.marmitt.core.ports.outbound.exchange.adapter.SenderSpecializedProcessorPort;
 
-public class CancelOrderProcessor implements SenderSpecializedProcessorPort {
+
+public class CancelOrderProcessor implements CoinbaseSenderProcessor {
 
     private final ObjectMapper objectMapper;
