+++ adapter-coinbase/src/main/java/com/marmitt/coinbase/processor/send/CoinbaseSenderMessageProcessor.java
@@ -3,7 +3,7 @@ package com.marmitt.coinbase.processor.send;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.marmitt.core.dto.websocket.request.MessageRequest;
 import com.marmitt.core.ports.outbound.exchange.adapter.SenderMessageProcessorPort;
-import com.marmitt.core.ports.outbound.exchange.adapter.SenderSpecializedProcessorPort;
+
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.ArrayList;
@@ -12,7 +12,7 @@ import java.util.List;
 @Slf4j
 public class CoinbaseSenderMessageProcessor implements SenderMessageProcessorPort {
 
-    private final List<SenderSpecializedProcessorPort> specializedProcessors;
+    private final List<CoinbaseSenderProcessor> specializedProcessors;
 
     public CoinbaseSenderMessageProcessor(ObjectMapper objectMapper) {
         this.specializedProcessors = new ArrayList<>();
