+++ adapter-coinbase/src/main/java/com/marmitt/coinbase/processor/send/CoinbaseSenderProcessor.java
@@ -1,0 +1,11 @@
+package com.marmitt.coinbase.processor.send;
+
+import com.marmitt.core.dto.websocket.request.MessageRequest;
+import com.marmitt.core.enums.MessageType;
+
+interface CoinbaseSenderProcessor {
+
+    String execute(MessageRequest request);
+
+    boolean canProcess(MessageType messageType);
+}
